requirements.txt
Django==1.9
Pillow==5.1.0
Pillow-PIL==0.1.dev0
psycopg2==2.7.4
